Refactor: pull some UI into ui.py
